vulkaninfo: Fix issue 202
Vulkaninfo was replacing function `snprintf`
with function `_snprintf` on windows.

This seems to be for backwards compatiblity
with VS2013.

`_snprintf` is considered deprecated in VS2015
as `snprintf` is properly implemented.

Since we are dropping compatibility with VS2013 it
seems like this replacement can be removed.

Thus, this commit removes the pre-processor command
that replaces `snprintf` with `_snprintf`.

Change-Id: I85f726fcb0a1cfcc902487100d35dc63d2ba00d8